Update and rename TowerHanoi.cpp to TowerOfHanoi.cpp

+++ puzzles/Tower_Of_Hanoi/c++/TowerOfHanoi.cpp
@@ -1,7 +1,7 @@
 #include<bits/stdc++.h>
 using namespace std;
 
-void move(int n,char pole1,char pole2,char pole3)
+void move(long long n,char pole1,char pole2,char pole3)
 {
 	if(n==0) return;
 	move(n-1,pole1,pole3,pole2);
@@ -13,7 +13,8 @@ int main()
 	cin.tie(NULL);
 	cout.tie(NULL);
 	ios_base::sync_with_stdio(false);
-	int n;
+	
+	long long n;
 	cin>>n;
 	move(n,'a','b','c');
 }
